refactor(network): replace NetworkErrors with Throw helper
